+++ docs/source/changelog.md
@@ -8,3 +8,4 @@
 - All command line programs now has `-a`, `--apply` option which writes the refactored version back to the file instead of dumping the diff.
 - Some of non-deterministic refactors (generating variations of the same code) are now early exitted.
 - `refactor.context.ScopeInfo` objects now can list definitions made in their context.
+- `refactor.context.Context` objects now store the path for the module they are processing if it is available.